New version 0.5.0. Fix data type

+++ client/src/main/kotlin/com/rarible/opensea/client/model/OpenSeaOrder.kt
@@ -104,7 +104,7 @@ data class OpenSeaOrder(
 
     val quantity: BigInteger,
 
-    val v: BigInteger,
+    val v: Byte,
 
     val r: Binary,
